pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "travel-multi-agent-planner"
version = "0.1.0"
description = "A multi-agent travel planning system for an Advanced AI course project."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.37",
    "pandas>=2.2",
    "requests>=2.32",
    "python-dotenv>=1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["travel_multi_agent_planner*"]
